refactor: useEvaluationData.js 분리
